Now displaying profile name as title
User's profile names are now used as the title and heading for their LittleLink pages. Previously the '@name' was used for this, so for example the page name x.com/@admin is now not tied to the @admin part anymore, so a space could be used in the name for example.

See an example of this here: https://i.imgur.com/iBS8Hsm.png