Added custom color for custom icon
This addition is mega janky. Essentially, if a color code (e.g. white; or #FFFFFF;) is entered before the custom icon, this will change the custom icon's color. Essentially, this just pastes the custom color database entry as a style element on the icon.

I will use this until I find a proper solution for this problem.

For custom icons:
See: https://blog.littlelink-custom.com/upcoming-features/
And: https://blog.littlelink-custom.com/progress-of-the-new-button-editor/